requirements.txt
pyautogen==0.2.34
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.30.0
pytest>=8.0.0
streamlit>=1.32.0